It now loads Groups, lets users post and comment on posts, and report.
Reporting is broken, though. It does not get the information required for the post. The API is not being used correctly. Users cannot join groups, and no users are shown to be members. Slashes and apostrophes (for example) are not formatted correctly. Usernames are not shown, settings are not available, messages and notifications do not load, and you cannot do anything that involves actually addressing another user.